A - Batting Average
Problem Statement
Takahashi is making a computer baseball game.
He will write a program that shows a batter's batting average with a specified number of digits.

There are integers A and B, which satisfy 1≤A≤10 and 0≤B≤A.
Let S be the string obtained as follows.

Round off  
A
B
​
  to three decimal digits, then write the integer part (1 digit), . (the decimal point), and the decimal part (3 digits) in this order, with trailing zeros.
For example, if A=7 and B=4, then  
A
B
​
 = 
7
4
​
 =0.571428… rounded off to three decimal digits is 0.571. Thus, S is 0.571.

You are given A and B as the input and asked to print S.

Constraints
1≤A≤10
0≤B≤A
A and B are integers.
Input
The input is given from Standard Input in the following format:

A B
Output
Print S in the format specified in the Problem Statement. Note that answers in different formats will be considered wrong.