A - Find Takahashi
Problem Statement
There are N bridges in AtCoder Village. The height of the bridge numbered i is Hi(i is an integer between 1 and N).
Every two different bridges in the village have different heights.

Print the number representing the highest bridge in the village.

Constraints
1≤N≤100
1≤Hi≤10^9
 
All Hi are different.
All values in the input are integers.