B - Let's Get a Perfect Score
N participants, numbered 1 to N, will participate in a contest with M problems, numbered 1 to M.

For an integer i between 1 and N and an integer j between 1 and M, participant i can solve problem j if the j-th character of Si is o, and cannot solve it if that character is x.

The participants must be in pairs. Print the number of ways to form a pair of participants who can collectively solve all the M problems.

More formally, print the number of pairs (x,y) of integers satisfying 1≤x<y≤N such that for any integer j between 1 and M, at least one of participant x and participant y can solve problem j.

Constraints
N is an integer between 2 and 30, inclusive.
M is an integer between 1 and 30, inclusive.
Si is a string of length M consisting of o and x.