C - String Delimiter
You are given a string S of length N consisting of lowercase English letters, ,, and ". It is guaranteed that S contains an even number of ".

Let 2K be the number of " in S. For each i=1,2,…,K, the characters from the (2i−1)-th " through the (2i)-th " are said to be enclosed.

Your task is to replace each , in S that is not an enclosed character with . and print the resulting string.

Constraints
N is an integer between 1 and 2×10^5 , inclusive.
S is a string of length N consisting of lowercase English letters, ,, and ".
S contains an even number of ".